backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
numpy==1.26.4
Pillow==9.0.1